+++ UnleashedRecomp/os/linux/logger_linux.cpp
@@ -12,6 +12,10 @@ void os::logger::SetWatchdogSuspended(bool suspended)
 {
 }
 
+void os::logger::SetTestLocation(const std::string_view location)
+{
+}
+
 void os::logger::Log(const std::string_view str, ELogType type, const char* func)
 {
     if (func)