+++ UnleashedRecomp/os/win32/logger_win32.cpp
@@ -19,6 +19,10 @@ void os::logger::SetWatchdogSuspended(bool suspended)
 {
 }
 
+void os::logger::SetTestLocation(const std::string_view location)
+{
+}
+
 void os::logger::Log(const std::string_view str, ELogType type, const char* func)
 {
     if (!os::process::g_consoleVisible)